api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "docxtpl>=0.18.0",
    "fastapi>=0.133.0",
    "firebase-admin>=7.1.0",
    "google-cloud-storage>=3.9.0",
    "google-cloud-tasks>=2.18.0",
    "jsonschema>=4.26.0",
    "langchain-core>=1.2.16",
    "langchain-openai>=1.1.10",
    "langsmith>=0.7.9",
    "openai>=2.24.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "uvicorn>=0.41.0",
]

[dependency-groups]
dev = [
    "ruff>=0.15.2",
]

[tool.ruff]
line-length = 100
target-version = "py313"
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "*.pyc",
    ".pytest_cache",
    ".mypy_cache",
    "build",
    "dist",
]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "UP",     # pyupgrade
    "RUF",    # Ruff-specific rules
]
ignore = [
    "E501",   # line too long (handled by formatter)
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # Allow unused imports in __init__ files

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
